Move PGPServiceBridge to BaseApplication
