+++ src/main/java/co/gounplugged/unpluggeddroid/activities/ChatActivity.java
@@ -52,24 +52,6 @@ public class ChatActivity extends BaseActivity {
     private ConversationContainer mConversationContainer;
     private ImageView mImageViewDropZoneChats, mImageViewDropZoneDelete;
 
-    private OpenPGPBridgeService mOpenPGPBridgeService;
-    private ServiceConnection mOpenPGPBridgeConnection = new ServiceConnection() {
-        @Override
-        public void onServiceConnected(ComponentName name, IBinder service) {
-            OpenPGPBridgeService.LocalBinder binder = (OpenPGPBridgeService.LocalBinder) service;
-            mOpenPGPBridgeService = binder.getService();
-            mIsBoundToOpenPGP = true;
-            Log.d(TAG, "bound to pgp bridge");
-        }
-
-        @Override
-        public void onServiceDisconnected(ComponentName name) {
-            mIsBoundToOpenPGP = false;
-            Log.d(TAG, "unbound from pgp bridge");
-        }
-    };;
-    private boolean mIsBoundToOpenPGP = false;
-
     private Conversation mSelectedConversation;
     private Conversation mClickedConversation;  //TODO refactor global var
 
@@ -130,11 +112,6 @@ public class ChatActivity extends BaseActivity {
 
         Log.d(TAG, "onStart");
 
-        bindService(
-                new Intent(this, OpenPGPBridgeService.class),
-                mOpenPGPBridgeConnection,
-                Context.BIND_AUTO_CREATE);
-
         ((BaseApplication) getApplicationContext()).seedKnownMasks();
     }
 
@@ -143,10 +120,6 @@ public class ChatActivity extends BaseActivity {
         super.onStop();
         Log.d(TAG, "onStop");
 
-        if (mIsBoundToOpenPGP) {
-            unbindService(mOpenPGPBridgeConnection);
-            mIsBoundToOpenPGP = false;
-        }
     }
 
     @Override
@@ -322,10 +295,6 @@ public class ChatActivity extends BaseActivity {
         if(mSelectedConversation == null) return true;
         if(mSelectedConversation.equals(newConversation)) return false;
         return true;
-    }
-
-    public OpenPGPBridgeService getOpenPGPBridgeService() {
-        return mOpenPGPBridgeService;
     }
 
     private class FragmentPagerAdapter extends android.support.v4.app.FragmentPagerAdapter {
