+++ src/main/java/co/gounplugged/unpluggeddroid/exceptions/EncryptionUnavailableException.java
@@ -1,0 +1,10 @@
+package co.gounplugged.unpluggeddroid.exceptions;
+
+/**
+ * Created by Marvin Arnold on 16/06/15.
+ */
+public class EncryptionUnavailableException extends Exception {
+    public EncryptionUnavailableException(String message) {
+        super(message);
+    }
+}
